Remove rosnode depency of GaitGeneratorController tests
