+++ march_rqt_gait_generator/package.xml
@@ -22,7 +22,6 @@
   <exec_depend>python-mock</exec_depend>
   <exec_depend>urdfdom_py</exec_depend>
 
-  <test_depend>rostest</test_depend>
   <test_depend>rosunit</test_depend>
 
   <export>
