--- march_rqt_gait_generator/test/launch/testing_gait_generator_controller.test
@@ -1,8 +1,0 @@
-<launch>
-    <arg name="robot" default="march4" doc="The robot to run. Can be: march3, march4, test_joint_rotational."/>
-
-    <param name="robot_description"
-           command="$(find xacro)/xacro '$(find march_description)/urdf/$(arg robot).xacro' "/>
-
-    <test  test-name="test_gait_generator_controller" pkg="march_rqt_gait_generator" type="gait_generator_controller_test"/>
-</launch>
